Split JS classes
